Latest Commit: Implemented separate header files for Material and Hitrecord
